+++ lib/src/generator/generate.dart
@@ -1,0 +1,22 @@
+import 'package:fast_i18n/src/generator/generate_header.dart';
+import 'package:fast_i18n/src/generator/generate_translations.dart';
+import 'package:fast_i18n/src/model/i18n_config.dart';
+import 'package:fast_i18n/src/model/i18n_data.dart';
+
+/// main generate function
+/// returns a string representing the content of the .g.dart file
+String generate(
+    {required I18nConfig config, required List<I18nData> translations}) {
+  StringBuffer buffer = StringBuffer();
+
+  generateHeader(buffer, config, translations);
+
+  buffer.writeln();
+  buffer.writeln('// translations');
+
+  for (I18nData localeData in translations) {
+    generateLocale(buffer, config, localeData);
+  }
+
+  return buffer.toString();
+}