+++ lib/src/parser_json.dart
@@ -1,6 +1,9 @@
 import 'dart:convert';
 
-import 'package:fast_i18n/src/model.dart';
+import 'package:fast_i18n/src/model/i18n_config.dart';
+import 'package:fast_i18n/src/model/i18n_data.dart';
+import 'package:fast_i18n/src/model/i18n_locale.dart';
+import 'package:fast_i18n/src/model/node.dart';
 
 /// parses a json of one locale
 /// returns an I18nData object