+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 4.0.1
+## 4.1.0-dev.0
 
 A rebuild is needed: `flutter pub run build_runner build`.